pyproject.toml
[tool.poetry]
name = "repl_python3_beta_TokenRevoker"
version = "0.1.0"
description = ""
authors = ["CoolCoderSJ <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^2.0.0"
PyGithub = "^1.55"
ipy = "^1.01"
easypydb = "^0.4.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
